@@ -1,92 +1,87 @@
-﻿using System;
+﻿using GTPS2ModelTool.Core;
+
+using PDTools.Files.Models.PS2.Commands;
+
+using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Xml.Linq;
 
-namespace PDTools.Files.Models.PS2.ModelSet
+namespace PDTools.Files.Models.PS2.ModelSet;
+
+public class ModelCommandShapeExtractor
 {
-    public class ModelCommandContext
+    public RenderCommandContext RenderCommandContext { get; set; } = new();
+
+    public string ModelName { get; set; }
+    public string ExtraShapeName { get; set; }
+
+    public int ModelIndex { get; set; }
+    public int CurrentLOD { get; set; } = -1;
+
+    public List<DumpedLOD> LODToShapes { get; set; } = [];
+
+    private int _texSetIndex { get; set; }
+
+    public string OutputDir { get; set; }
+
+    public ModelCallbackParameter? CurrentCallback { get; set; }
+    public int CallbackBranchIndex { get; set; } = -1;
+
+    public List<ModelSetupPS2Command> CommandQueue { get; set; } = [];
+
+    public void SetTexTable(int index)
     {
-        public int CallbackParameter { get; set; } = -1;
-        public string ExtraName { get; set; }
+        _texSetIndex = index;
+    }
 
-        public int ModelIndex { get; set; }
-        public int LOD { get; set; } = -1;
+    public void SetLOD(int lod)
+    {
+        CurrentLOD = lod;
 
-        private int _texSetIndex { get; set; }
+        if (LODToShapes.Count - 1 < lod)
+            LODToShapes.Add(new DumpedLOD());
+    }
 
-        public string OutputDir { get; set; }
-        public StreamWriter ObjWriter { get; set; }
-        public StreamWriter MatWriter { get; set; }
-
-        public int VertIdxStart { get; set; }
-        public int VTIdxStart { get; set; }
-
-        public void SetTexTable(int index)
+    public void AddShape(string name, PGLUshapeConverted shape)
+    {
+        if (CurrentLOD == -1)
         {
-            _texSetIndex = index;
+            CurrentLOD = 0;
+            LODToShapes.Add(new DumpedLOD());
         }
 
-        public void SetLOD(int lod)
+        shape.TextureSetIndex = _texSetIndex;
+        LODToShapes[CurrentLOD].Add(name, shape);
+
+        if (CurrentCallback is not null)
         {
-            LOD = lod;
-            ResetIndices();
+            LODToShapes[CurrentLOD].AddCallbackShape(CurrentCallback.Value, CallbackBranchIndex, name, shape);
         }
+    }
+}
 
-        public void ResetIndices()
-        {
-            VertIdxStart = 0;
-            VTIdxStart = 0;
-        }
+public class DumpedLOD
+{
+    public Dictionary<string, PGLUshapeConverted> Shapes { get; set; } = [];
+    public Dictionary<ModelCallbackParameter, List<List<string>>> Callbacks { get; set; } = [];
 
-        public void DumpShapeToObj(PGLUshapeConverted shape, int shapeIndex, string shapeName)
-        {
-            HashSet<int> texIndices = new HashSet<int>();
-            for (int i = 0; i < shape.Faces.Count; i++)
-            {
-                if (!texIndices.Contains(shape.Faces[i].TexId))
-                    texIndices.Add(shape.Faces[i].TexId);
-            }
+    public void Add(string name, PGLUshapeConverted shape)
+    {
+        if (!Shapes.ContainsKey(name)) // sometimes a shape can be called again? it's weird, needs investigation (gt4: ac_0014, model index 0)
+            Shapes.Add(name, shape);
+    }
 
-            bool isExternalTextureReflection = texIndices.All(e => e == VIFDescriptor.EXTERNAL_TEXTURE);
+    public void AddCallbackShape(ModelCallbackParameter parameter, int idx, string name, PGLUshapeConverted shape)
+    {
+        Callbacks.TryAdd(parameter, []);
 
-            ObjWriter.WriteLine($"# Shape {shapeIndex}");
+        while (Callbacks[parameter].Count <= idx)
+            Callbacks[parameter].Add([]);
 
-            if (isExternalTextureReflection)
-                ObjWriter.WriteLine($"o {shapeName}_reflection");
-            else
-                ObjWriter.WriteLine($"o {shapeName}");
-
-            shape.Dump(ObjWriter, MatWriter, _texSetIndex, VertIdxStart, VTIdxStart);
-            ObjWriter.WriteLine();
-
-            VertIdxStart += shape.Vertices.Count;
-            VTIdxStart += shape.UVs.Count;
-        }
-
-        public void SetupObjWriter(string name)
-        {
-            ObjWriter = new StreamWriter(Path.Combine(OutputDir, $"{name}.obj"));
-            MatWriter = new StreamWriter(Path.Combine(OutputDir, $"{name}.mtl"));
-        }
-
-        public void FinishModel()
-        {
-            if (ObjWriter is not null)
-            {
-                ObjWriter.Flush();
-                ObjWriter.Dispose();
-                ObjWriter = null;
-            }
-
-            if (MatWriter is not null)
-            {
-                MatWriter.Flush();
-                MatWriter.Dispose();
-                MatWriter = null;
-            }
-        }
+        Callbacks[parameter][idx].Add(name);
     }
 }
