@@ -13,424 +13,358 @@ using PDTools.Files.Models.PS2.Commands;
 using PDTools.Files.Models.VM;
 using PDTools.Files.Textures;
 
-namespace PDTools.Files.Models.PS2.ModelSet
+namespace PDTools.Files.Models.PS2.ModelSet;
+
+/// <summary>
+/// Model Set 2. Used by GT4
+/// </summary>
+public class ModelSet2 : ModelSetPS2Base
 {
     /// <summary>
-    /// Model Set 2. Used by GT4
+    /// "MDLS"
     /// </summary>
-    public class ModelSet2 : ModelSetPS2Base
+    public const uint MAGIC = 0x534C444D;
+    public const uint HeaderSize = 0x80;
+
+    public RelocatorBase Relocator { get; set; }
+
+    public byte InstanceFlags { get; set; }
+    public ushort InstanceOutRegisterCount { get; set; }
+    public ushort InstanceUnkRegisterCount { get; set; }
+    public ushort HostMethodInfoCount { get; set; }
+
+    public List<PGLUmaterial> Materials { get; set; } = [];
+    public List<List<PGLUmaterial>> VariationMaterials { get; set; } = [];
+
+    public List<RegisterInfo> OutRegisterInfos { get; set; } = [];
+    public List<RegisterInfo> ExternalInfos { get; set; } = [];
+    public List<RegisterInfo> HostMethodInfos { get; set; } = [];
+
+    /// <summary>
+    /// Textures. Each texture set within a list represents seemingly one lod level.
+    /// </summary>
+    public List<List<TextureSet1>> TextureSetLists { get; set; } = [];
+
+    public void FromStream(Stream stream)
     {
-        /// <summary>
-        /// "MDLS"
-        /// </summary>
-        public const uint MAGIC = 0x534C444D;
-        public const uint HeaderSize = 0x80;
+        var bs = new BinaryStream(stream);
+        long basePos = bs.Position;
 
-        public RelocatorBase Relocator { get; set; }
+        uint magic = bs.ReadUInt32();
+        if (magic != MAGIC)
+            throw new InvalidDataException("Not a valid ModelSet2 stream.");
 
-        public byte InstanceFlags { get; set; }
-        public ushort InstanceOutRegisterCount { get; set; }
-        public ushort InstanceUnkRegisterCount { get; set; }
-        public ushort HostMethodInfoCount { get; set; }
+        /* HEADER - 0xE4 */
+        int relocatorInfoOffset = bs.ReadInt32();
+        int relocatorDataSize = bs.ReadInt32();
+        int relocationBase = bs.ReadInt32();
+        int fileSize = bs.ReadInt32();
 
-        public List<ModelSet2Model> Models { get; set; } = new List<ModelSet2Model>();
-        public List<PGLUshape> Shapes { get; set; } = new List<PGLUshape>();
-        public List<PGLUmaterial> Materials { get; set; } = new List<PGLUmaterial>();
-        public List<List<PGLUmaterial>> VariationMaterials { get; set; } = new List<List<PGLUmaterial>>();
+        byte unk = bs.Read1Byte();
+        InstanceFlags = bs.Read1Byte();
 
-        public List<RegisterInfo> OutRegisterInfos { get; set; } = new List<RegisterInfo>();
-        public List<RegisterInfo> ExternalInfos { get; set; } = new List<RegisterInfo>();
-        public List<RegisterInfo> HostMethodInfos { get; set; } = new List<RegisterInfo>();
+        // Counts
+        ushort modelCount = bs.ReadUInt16();
+        ushort shapeCount = bs.ReadUInt16();
+        ushort materialCount = bs.ReadUInt16();
+        ushort texSetCount = bs.ReadUInt16();
+        ushort colorCount = bs.ReadUInt16();
+        InstanceOutRegisterCount = bs.ReadUInt16();
+        InstanceUnkRegisterCount = bs.ReadUInt16();
+        HostMethodInfoCount = bs.ReadUInt16();
+        ushort externalInfoCount = bs.ReadUInt16();
+        ushort outRegisterInfoCount = bs.ReadUInt16();
+        ushort symbolsCount = bs.ReadUInt16();
+        byte variationMaterialCount = bs.Read1Byte();
+        byte currentColorIndex = bs.Read1Byte();
+        ushort bindMatrixCount = bs.ReadUInt16();
+        bs.Position += 0x06;
 
-        /// <summary>
-        /// Textures. Each texture set within a list represents seemingly one lod level.
-        /// </summary>
-        public List<List<TextureSet1>> TextureSetLists { get; set; } = new List<List<TextureSet1>>();
+        ushort instanceSize = bs.ReadUInt16();
 
-        public void FromStream(Stream stream)
+        // Offsets
+        uint modelsOffset = bs.ReadUInt32();
+        uint shapesOffset = bs.ReadUInt32();
+        uint materialsOffset = bs.ReadUInt32();
+        uint pgluTexSetsOffset = bs.ReadUInt32();
+        uint bindMatricesOffset = bs.ReadUInt32();
+        uint hostMethodInfosOffset = bs.ReadUInt32();
+        uint externalInfoOffset = bs.ReadUInt32();
+        uint outRegisterInfoOffset = bs.ReadUInt32();
+        uint symbolsOffset = bs.ReadUInt32();
+        uint vmBytecodeOffset = bs.ReadUInt32();
+        uint variationMaterialsOffset = bs.ReadUInt32();
+
+        bs.Position = basePos + shapesOffset;
+
+        bs.Position = basePos + relocatorInfoOffset;
+        Relocator = RelocatorBase.FromStream(bs, basePos);
+        Relocator.MakeRelocatableGroups();
+
+        ReadMaterials(bs, basePos, materialsOffset, materialCount);
+        ReadVariationMaterials(bs, basePos, variationMaterialsOffset, variationMaterialCount);
+        ReadModels(bs, basePos, modelsOffset, modelCount);
+        ReadShapes(bs, basePos, shapesOffset, shapeCount);
+        ReadTextureSets(bs, basePos, pgluTexSetsOffset, colorCount, texSetCount);
+        ReadHostMethodInfos(bs, basePos, hostMethodInfosOffset, HostMethodInfoCount);
+        ReadExternalInfos(bs, basePos, externalInfoOffset, externalInfoCount);
+        ReadOutRegisterInfo(bs, basePos, outRegisterInfoOffset, outRegisterInfoCount);
+
+        Instance instance = new Instance();
+        instance.ModelSet = this;
+
+        if ((InstanceFlags & 1) != 0)
         {
-            using var bs = new BinaryStream(stream);
-            long basePos = bs.Position;
-
-            uint magic = bs.ReadUInt32();
-            if (magic != MAGIC)
-                throw new InvalidDataException("Not a valid ModelSet2 stream.");
-
-            /* HEADER - 0xE4 */
-            int relocatorInfoOffset = bs.ReadInt32();
-            int relocatorDataSize = bs.ReadInt32();
-            int relocationBase = bs.ReadInt32();
-            int fileSize = bs.ReadInt32();
-
-            byte unk = bs.Read1Byte();
-            InstanceFlags = bs.Read1Byte();
-
-            // Counts
-            ushort modelCount = bs.ReadUInt16();
-            ushort shapeCount = bs.ReadUInt16();
-            ushort materialCount = bs.ReadUInt16();
-            ushort texSetCount = bs.ReadUInt16();
-            ushort colorCount = bs.ReadUInt16();
-            InstanceOutRegisterCount = bs.ReadUInt16();
-            InstanceUnkRegisterCount = bs.ReadUInt16();
-            HostMethodInfoCount = bs.ReadUInt16();
-            ushort externalInfoCount = bs.ReadUInt16();
-            ushort outRegisterInfoCount = bs.ReadUInt16();
-            ushort symbolsCount = bs.ReadUInt16();
-            byte variationMaterialCount = bs.Read1Byte();
-            byte currentColorIndex = bs.Read1Byte();
-            ushort bindMatrixCount = bs.ReadUInt16();
-            bs.Position += 0x06;
-
-            ushort instanceSize = bs.ReadUInt16();
-
-            // Offsets
-            uint modelsOffset = bs.ReadUInt32();
-            uint shapesOffset = bs.ReadUInt32();
-            uint materialsOffset = bs.ReadUInt32();
-            uint pgluTexSetsOffset = bs.ReadUInt32();
-            uint bindMatricesOffset = bs.ReadUInt32();
-            uint hostMethodInfosOffset = bs.ReadUInt32();
-            uint externalInfoOffset = bs.ReadUInt32();
-            uint outRegisterInfoOffset = bs.ReadUInt32();
-            uint symbolsOffset = bs.ReadUInt32();
-            uint vmBytecodeOffset = bs.ReadUInt32();
-            uint variationMaterialsOffset = bs.ReadUInt32();
-
-            bs.Position = basePos + shapesOffset;
-
-            bs.Position = basePos + relocatorInfoOffset;
-            Relocator = RelocatorBase.FromStream(bs, basePos);
-            Relocator.MakeRelocatableGroups();
-
-            ReadMaterials(bs, basePos, materialsOffset, materialCount);
-            ReadVariationMaterials(bs, basePos, variationMaterialsOffset, variationMaterialCount);
-            ReadModels(bs, basePos, modelsOffset, modelCount);
-            ReadShapes(bs, basePos, shapesOffset, shapeCount);
-            ReadTextureSets(bs, basePos, pgluTexSetsOffset, colorCount, texSetCount);
-            ReadHostMethodInfos(bs, basePos, hostMethodInfosOffset, HostMethodInfoCount);
-            ReadExternalInfos(bs, basePos, externalInfoOffset, externalInfoCount);
-            ReadOutRegisterInfo(bs, basePos, outRegisterInfoOffset, outRegisterInfoCount);
-
-            Instance instance = new Instance();
-            instance.ModelSet = this;
-
-            if ((InstanceFlags & 1) != 0)
-            {
-                bs.Position = basePos + 0x7C;
-                uint instanceOffset = bs.ReadUInt32();
-                bs.Position = instanceOffset;
-                instance.Read(bs, basePos);
-            }
-            else
-            {
-                instance.OutputRegisters = new RegisterVal[InstanceOutRegisterCount];
-                instance.Unk2 = new RegisterVal[InstanceUnkRegisterCount];
-                instance.HostMethodRegisters = new RegisterVal[HostMethodInfoCount];
-            }
-
+            bs.Position = basePos + 0x7C;
+            uint instanceOffset = bs.ReadUInt32();
+            bs.Position = instanceOffset;
+            instance.Read(bs, basePos);
+        }
+        else
+        {
+            instance.OutputRegisters = new RegisterVal[InstanceOutRegisterCount];
+            instance.Unk2 = new RegisterVal[InstanceUnkRegisterCount];
+            instance.HostMethodRegisters = new RegisterVal[HostMethodInfoCount];
         }
 
-        private void ReadMaterials(BinaryStream bs, long baseMdlPos, uint offset, uint count)
-        {
-            for (ushort i = 0; i < count; i++)
-            {
-                bs.Position = baseMdlPos + offset + i * PGLUmaterial.GetSize();
+    }
 
+    private void ReadMaterials(BinaryStream bs, long baseMdlPos, uint offset, uint count)
+    {
+        for (ushort i = 0; i < count; i++)
+        {
+            bs.Position = baseMdlPos + offset + i * PGLUmaterial.GetSize();
+
+            var material = new PGLUmaterial();
+            material.FromStream(bs, baseMdlPos);
+            Materials.Add(material);
+        }
+    }
+
+    private void ReadVariationMaterials(BinaryStream bs, long baseMdlPos, uint offset, uint count)
+    {
+        for (ushort i = 0; i < count; i++)
+        {
+            bs.Position = baseMdlPos + offset + (i * 4);
+
+            uint materialsOffset = bs.ReadUInt32();
+            bs.Position = baseMdlPos + materialsOffset;
+            VariationMaterials.Add([]);
+
+            for (int j = 0; j < Materials.Count; j++)
+            {
                 var material = new PGLUmaterial();
                 material.FromStream(bs, baseMdlPos);
-                Materials.Add(material);
+                VariationMaterials[i].Add(material);
             }
         }
+    }
 
-        private void ReadVariationMaterials(BinaryStream bs, long baseMdlPos, uint offset, uint count)
+    private void ReadModels(BinaryStream bs, long baseMdlPos, uint offset, uint count)
+    {
+        for (ushort i = 0; i < count; i++)
         {
-            for (ushort i = 0; i < count; i++)
-            {
-                bs.Position = baseMdlPos + offset + (i * 4);
+            bs.Position = baseMdlPos + offset + i * ModelSet2Model.GetSize();
 
-                uint materialsOffset = bs.ReadUInt32();
-                bs.Position = baseMdlPos + materialsOffset;
-                VariationMaterials.Add(new List<PGLUmaterial>());
-
-                for (int j = 0; j < Materials.Count; j++)
-                {
-                    var material = new PGLUmaterial();
-                    material.FromStream(bs, baseMdlPos);
-                    VariationMaterials[i].Add(material);
-                }
-            }
+            var model = new ModelSet2Model();
+            model.FromStream(bs, baseMdlPos);
+            Models.Add(model);
         }
+    }
 
-        private void ReadModels(BinaryStream bs, long baseMdlPos, uint offset, uint count)
+    private void ReadShapes(BinaryStream bs, long baseMdlPos, uint offset, uint count)
+    {
+        for (ushort i = 0; i < count; i++)
         {
-            for (ushort i = 0; i < count; i++)
-            {
-                bs.Position = baseMdlPos + offset + i * ModelSet2Model.GetSize();
+            bs.Position = baseMdlPos + offset + i * sizeof(int);
+            uint shapeOffset = bs.ReadUInt32();
 
-                var model = new ModelSet2Model();
-                model.FromStream(bs, baseMdlPos);
-                Models.Add(model);
-            }
-        }
+            bs.Position = baseMdlPos + shapeOffset;
 
-        private void ReadShapes(BinaryStream bs, long baseMdlPos, uint offset, uint count)
-        {
-            for (ushort i = 0; i < count; i++)
-            {
-                bs.Position = baseMdlPos + offset + i * sizeof(int);
-                uint shapeOffset = bs.ReadUInt32();
-
-                bs.Position = baseMdlPos + shapeOffset;
-
-                var shape = new PGLUshape();
-                shape.FromStream(bs, baseMdlPos);
-                Shapes.Add(shape);
-            }
-        }
-
-        private void ReadHostMethodInfos(BinaryStream bs, long baseMdlPos, uint offset, uint count)
-        {
-            for (ushort i = 0; i < count; i++)
-            {
-                bs.Position = baseMdlPos + offset + i * RegisterInfo.GetSize();
-
-                RegisterInfo registerInfo = new RegisterInfo();
-                registerInfo.FromStream(bs, baseMdlPos);
-                HostMethodInfos.Add(registerInfo);
-            }
-        }
-
-        private void ReadExternalInfos(BinaryStream bs, long baseMdlPos, uint offset, uint count)
-        {
-            for (ushort i = 0; i < count; i++)
-            {
-                bs.Position = baseMdlPos + offset + i * RegisterInfo.GetSize();
-
-                RegisterInfo registerInfo = new RegisterInfo();
-                registerInfo.FromStream(bs, baseMdlPos);
-                ExternalInfos.Add(registerInfo);
-            }
-        }
-
-        private void ReadOutRegisterInfo(BinaryStream bs, long baseMdlPos, uint offset, uint count)
-        {
-            for (ushort i = 0; i < count; i++)
-            {
-                bs.Position = baseMdlPos + offset + i * RegisterInfo.GetSize();
-
-                RegisterInfo registerInfo = new RegisterInfo();
-                registerInfo.FromStream(bs, baseMdlPos);
-                OutRegisterInfos.Add(registerInfo);
-            }
-        }
-
-        private void ReadTextureSets(BinaryStream bs, long baseMdlPos, uint offset, uint listCount, uint textureLodLevels)
-        {
-            for (int i = 0; i < listCount; i++)
-            {
-                bs.Position = baseMdlPos + offset + i * sizeof(int);
-                int entriesOffset = bs.ReadInt32();
-
-                // One texture set per lod level
-                List<TextureSet1> list = new List<TextureSet1>();
-                for (int j = 0; j < textureLodLevels; j++)
-                {
-                    bs.Position = baseMdlPos + entriesOffset + j * sizeof(int);
-
-                    int off = bs.ReadInt32();
-                    bs.Position = baseMdlPos + off;
-
-                    TextureSet1 textureSet = new TextureSet1();
-                    textureSet.FromStream(bs);
-                    list.Add(textureSet);
-                }
-
-                TextureSetLists.Add(list);
-            }
-        }
-
-        public void DumpModel(int modelIndex, string outdir)
-        {
-            ModelCommandContext context = new ModelCommandContext();
-            context.ModelIndex = modelIndex;
-            context.OutputDir = outdir;
-
-            ModelSet2Model model = Models[modelIndex];
-
-            ProcessCommands(context, model.Commands);
-        }
-
-        private void ProcessCommands(ModelCommandContext context, List<ModelSetupPS2Command> cmds)
-        {
-            foreach (ModelSetupPS2Command command in cmds)
-            {
-                switch (command.Opcode)
-                {
-                    case ModelSetupPS2Opcode.BBoxRender:
-                        ProcessCommands(context, (command as Cmd_BBoxRender).CommandsOnRender);
-                        context.FinishModel();
-                        break;
-
-                    case ModelSetupPS2Opcode.LODSelect:
-                        var lodSel = command as Cmd_LODSelect;
-                        for (int i = 0; i < lodSel.CommandsPerLOD.Count; i++)
-                        {
-                            context.SetLOD(i);
-
-                            ProcessCommands(context, lodSel.CommandsPerLOD[i]);
-
-                            context.FinishModel();
-                        }
-                        break;
-
-                    case ModelSetupPS2Opcode.CallModelCallback:
-                        var callbackCmd = command as Cmd_CallModelCallback;
-                        if (callbackCmd.Parameter == 0) // Tail Lamp
-                        {
-                            ProcessCommands(context, callbackCmd.Default);
-
-                            context.ExtraName = "tail_lamp_off";
-                            ProcessCommands(context, callbackCmd.CommandsPerBranch[0]);
-
-                            context.ExtraName = "tail_lamp_on";
-                            ProcessCommands(context, callbackCmd.CommandsPerBranch[1]);
-
-                            context.ExtraName = null;
-                        }
-
-                        break;
-
-                    case ModelSetupPS2Opcode.pgluSetTexTable_Byte:
-                        byte index = (command as Cmd_pgluSetTexTable_Byte).TexSetTableIndex;
-                        context.SetTexTable(index);
-                        break;
-
-                    case ModelSetupPS2Opcode.pgluCallShape_Byte:
-
-                        if (context.ObjWriter is null)
-                        {
-                            if (context.LOD == -1)
-                            {
-                                context.SetupObjWriter($"model{context.ModelIndex}");
-                                context.ObjWriter.WriteLine($"mtllib model{context.ModelIndex}.mtl");
-                            }
-                            else
-                            {
-                                context.SetupObjWriter($"model{context.ModelIndex}.lod{context.LOD}");
-                                context.ObjWriter.WriteLine($"mtllib model{context.ModelIndex}.lod{context.LOD}.mtl");
-                            }
-
-                        }
-
-                        var callShape = (command as Cmd_pgluCallShapeByte);
-                        int shapeIndex = callShape.ShapeIndex;
-                        PGLUshapeConverted shapeData = Shapes[shapeIndex].GetShapeData();
-
-                        string name = $"shape{shapeIndex}";
-                        if (!string.IsNullOrEmpty(context.ExtraName))
-                            name += $"_{context.ExtraName}";
-
-                        context.DumpShapeToObj(shapeData, shapeIndex, name);
-                        break;
-                }
-            }
-        }
-
-        public override List<TextureSet1> GetTextureSetList()
-        {
-            if (TextureSetLists.Count == 0)
-                TextureSetLists.Add(new List<TextureSet1>());
-
-            return TextureSetLists[0];
-        }
-
-        public override int AddShape(PGLUshape shape)
-        {
+            var shape = new PGLUshape();
+            shape.FromStream(bs);
             Shapes.Add(shape);
-            return Shapes.Count - 1;
         }
+    }
 
-        public override int AddMaterial(PGLUmaterial material)
+    private void ReadHostMethodInfos(BinaryStream bs, long baseMdlPos, uint offset, uint count)
+    {
+        for (ushort i = 0; i < count; i++)
         {
-            Materials.Add(material);
-            return Materials.Count - 1;
+            bs.Position = baseMdlPos + offset + i * RegisterInfo.GetSize();
+
+            RegisterInfo registerInfo = new RegisterInfo();
+            registerInfo.FromStream(bs, baseMdlPos);
+            HostMethodInfos.Add(registerInfo);
         }
+    }
 
-        public override int GetMaterialCount()
+    private void ReadExternalInfos(BinaryStream bs, long baseMdlPos, uint offset, uint count)
+    {
+        for (ushort i = 0; i < count; i++)
         {
-            return Materials.Count;
+            bs.Position = baseMdlPos + offset + i * RegisterInfo.GetSize();
+
+            RegisterInfo registerInfo = new RegisterInfo();
+            registerInfo.FromStream(bs, baseMdlPos);
+            ExternalInfos.Add(registerInfo);
         }
+    }
 
-        public int GetInstanceSize()
+    private void ReadOutRegisterInfo(BinaryStream bs, long baseMdlPos, uint offset, uint count)
+    {
+        for (ushort i = 0; i < count; i++)
         {
-            int size = 0x20; // header/meta
-            size += (OutRegisterInfos.Count + InstanceUnkRegisterCount + HostMethodInfos.Count) * sizeof(uint);
+            bs.Position = baseMdlPos + offset + i * RegisterInfo.GetSize();
 
-            int maxRegisterSize = 0;
-            foreach (var hostMethodInfo in HostMethodInfos)
+            RegisterInfo registerInfo = new RegisterInfo();
+            registerInfo.FromStream(bs, baseMdlPos);
+            OutRegisterInfos.Add(registerInfo);
+        }
+    }
+
+    private void ReadTextureSets(BinaryStream bs, long baseMdlPos, uint offset, uint listCount, uint textureLodLevels)
+    {
+        for (int i = 0; i < listCount; i++)
+        {
+            bs.Position = baseMdlPos + offset + i * sizeof(int);
+            int entriesOffset = bs.ReadInt32();
+
+            // One texture set per lod level
+            List<TextureSet1> list = new();
+            for (int j = 0; j < textureLodLevels; j++)
             {
-                int registerSize = hostMethodInfo.RegisterIndex + hostMethodInfo.ArrayLength;
-                if (maxRegisterSize < registerSize)
-                    maxRegisterSize = registerSize;
+                bs.Position = baseMdlPos + entriesOffset + j * sizeof(int);
+
+                int off = bs.ReadInt32();
+                bs.Position = baseMdlPos + off;
+
+                TextureSet1 textureSet = new TextureSet1();
+                textureSet.FromStream(bs);
+                list.Add(textureSet);
             }
 
-            int totalRegisterSize = 0;
-            foreach (var hostMethodInfo in HostMethodInfos)
-            {
-                int registerSize = hostMethodInfo.RegisterIndex + hostMethodInfo.ArrayLength;
-                if (maxRegisterSize < registerSize)
-                    maxRegisterSize = registerSize;
-            }
+            TextureSetLists.Add(list);
+        }
+    }
 
-            size += totalRegisterSize * sizeof(int);
-            return size;
+    public override List<TextureSet1> GetTextureSetList()
+    {
+        if (TextureSetLists.Count == 0)
+            TextureSetLists.Add([]);
+
+        return TextureSetLists[0];
+    }
+
+    public override uint AddShape(PGLUshape shape)
+    {
+        Shapes.Add(shape);
+        return (uint)(Shapes.Count - 1);
+    }
+
+    public override int AddMaterial(PGLUmaterial material)
+    {
+        Materials.Add(material);
+        return Materials.Count - 1;
+    }
+
+    public override int GetMaterialCount()
+    {
+        return Materials.Count;
+    }
+
+    public int GetInstanceSize()
+    {
+        int size = 0x20; // header/meta
+        size += (OutRegisterInfos.Count + InstanceUnkRegisterCount + HostMethodInfos.Count) * sizeof(uint);
+
+        int maxRegisterSize = 0;
+        foreach (var hostMethodInfo in HostMethodInfos)
+        {
+            int registerSize = hostMethodInfo.RegisterIndex + hostMethodInfo.ArrayLength;
+            if (maxRegisterSize < registerSize)
+                maxRegisterSize = registerSize;
         }
 
-
-        public class Instance
+        int totalRegisterSize = 0;
+        foreach (var hostMethodInfo in HostMethodInfos)
         {
-            public ModelSet2 ModelSet { get; set; }
+            int registerSize = hostMethodInfo.RegisterIndex + hostMethodInfo.ArrayLength;
+            if (maxRegisterSize < registerSize)
+                maxRegisterSize = registerSize;
+        }
 
-            /// <summary>
-            /// Info about each register provided in model set out register infos
-            /// </summary>
-            public RegisterVal[] OutputRegisters { get; set; }
+        size += totalRegisterSize * sizeof(int);
+        return size;
+    }
 
-            /// <summary>
-            /// Maybe external?
-            /// </summary>
-            public RegisterVal[] Unk2 { get; set; }
+    public override List<PGLUmaterial> GetVariationMaterials(int varIndex)
+    {
+        return varIndex == 0 ? Materials : VariationMaterials[varIndex];
+    }
 
-            /// <summary>
-            /// Engine provided from the specified model set host method infos
-            /// </summary>
-            public RegisterVal[] HostMethodRegisters { get; set; }
+    public override int GetNumModels()
+    {
+        return Models.Count;
+    }
 
-            public void Read(BinaryStream bs, long baseMdlPos)
-            {
-                int parentModelSetPtr = bs.ReadInt32();
-                int outRegistersPtr = bs.ReadInt32();
-                int unkRegistersPtr = bs.ReadInt32();
-                int hostMethodRegistersPtr = bs.ReadInt32();
-                int hostMethodInfosFuncs = bs.ReadInt32();
-                // The rest we don't need
+    public override PGLUshape GetShape(int shapeIndex)
+    {
+        return Shapes[shapeIndex];
+    }
 
-                bs.Position = baseMdlPos + outRegistersPtr;
-                OutputRegisters = new RegisterVal[ModelSet.InstanceOutRegisterCount];
-                for (int i = 0; i < ModelSet.InstanceOutRegisterCount; i++)
-                    OutputRegisters[i] = new RegisterVal(bs.ReadInt32());
+    public override int GetNumShapes()
+    {
+        return Shapes.Count;
+    }
 
-                bs.Position = baseMdlPos + unkRegistersPtr;
-                Unk2 = new RegisterVal[ModelSet.InstanceUnkRegisterCount];
-                for (int i = 0; i < ModelSet.InstanceUnkRegisterCount; i++)
-                    Unk2[i] = new RegisterVal(bs.ReadInt32());
+    public override int GetNumVariations()
+    {
+        return Math.Max(1, VariationMaterials.Count);
+    }
 
-                bs.Position = baseMdlPos + hostMethodRegistersPtr;
-                HostMethodRegisters = new RegisterVal[ModelSet.HostMethodInfoCount];
-                for (int i = 0; i < ModelSet.HostMethodInfoCount; i++)
-                    HostMethodRegisters[i] = new RegisterVal(bs.ReadInt32());
-            }
+    public class Instance
+    {
+        public ModelSet2 ModelSet { get; set; }
+
+        /// <summary>
+        /// Info about each register provided in model set out register infos
+        /// </summary>
+        public RegisterVal[] OutputRegisters { get; set; }
+
+        /// <summary>
+        /// Maybe external?
+        /// </summary>
+        public RegisterVal[] Unk2 { get; set; }
+
+        /// <summary>
+        /// Engine provided from the specified model set host method infos
+        /// </summary>
+        public RegisterVal[] HostMethodRegisters { get; set; }
+
+        public void Read(BinaryStream bs, long baseMdlPos)
+        {
+            int parentModelSetPtr = bs.ReadInt32();
+            int outRegistersPtr = bs.ReadInt32();
+            int unkRegistersPtr = bs.ReadInt32();
+            int hostMethodRegistersPtr = bs.ReadInt32();
+            int hostMethodInfosFuncs = bs.ReadInt32();
+            // The rest we don't need
+
+            bs.Position = baseMdlPos + outRegistersPtr;
+            OutputRegisters = new RegisterVal[ModelSet.InstanceOutRegisterCount];
+            for (int i = 0; i < ModelSet.InstanceOutRegisterCount; i++)
+                OutputRegisters[i] = new RegisterVal(bs.ReadInt32());
+
+            bs.Position = baseMdlPos + unkRegistersPtr;
+            Unk2 = new RegisterVal[ModelSet.InstanceUnkRegisterCount];
+            for (int i = 0; i < ModelSet.InstanceUnkRegisterCount; i++)
+                Unk2[i] = new RegisterVal(bs.ReadInt32());
+
+            bs.Position = baseMdlPos + hostMethodRegistersPtr;
+            HostMethodRegisters = new RegisterVal[ModelSet.HostMethodInfoCount];
+            for (int i = 0; i < ModelSet.HostMethodInfoCount; i++)
+                HostMethodRegisters[i] = new RegisterVal(bs.ReadInt32());
         }
     }
 }
