@@ -10,11 +10,11 @@ using Syroot.BinaryData;
 namespace PDTools.Files.Models.PS2.Commands
 {
     /// <summary>
-    /// Same as command 54 but sets all 4 values to specified.
+    /// Same as command 50 but sets all 4 values to specified.
     /// </summary>
     public class Cmd_GT3_2_4f : ModelSetupPS2Command
     {
-        public override ModelSetupPS2Opcode Opcode => ModelSetupPS2Opcode.pglGT3_3_4f;
+        public override ModelSetupPS2Opcode Opcode => ModelSetupPS2Opcode.pglGT3_2_4f;
 
         // Color. RGB only, A is ignored
 
@@ -30,7 +30,10 @@ namespace PDTools.Files.Models.PS2.Commands
 
         public Cmd_GT3_2_4f(float r, float g, float b, float a)
         {
-
+            R = r;
+            G = g;
+            B = b;
+            A = a;
         }
 
         public override void Read(BinaryStream bs, int commandsBaseOffset)
